T21427 - make api calls compatible with our forked tornado
Summary:
tornado stores requests data in two different places:
- request body goes into self.body
- query string gets parsed out and put into self.arguments

I think the newer versions of tornado try to combine these but ours won't.

when we make a request with RequestsFetcher, the request data only shows up in self.body, and the url of the request is something like "/api/ping"

When we make a request with URLLibFetcher however, the request data shows up in the query string and the url is something like "/api/ping?worker=123"

since RPCHandler (the request handler which handles all interactions on the scheduler) only looks in self.request.arguments we should either
1) make it look in both arguments and body or
2) don't use RequestsFetcher

Test Plan:
Made the change locally and ran a small pipeline.

the error currently doesn't show up in prod because the luigi server is using
the most up-to-date version of tornado. To test this, I will install our version
of tornado and make the change locally and run both pipelines

Reviewers: adam, bill, jack

Reviewed By: jack

CC: samson

Maniphest Tasks: T21427

Differential Revision: https://phab.wish.com/D29321